Split off rust Urkel node marshaling into separate file
